refactor: Move unauthed group endpoints
