+++ src/routes/unauthenticated/mod.rs
@@ -1,1 +1,2 @@
 mod user;
+mod group;
